Add `TxBuilder` and `Signatory`
`TxBuilder` allows signing a tx.

`TxBuilder` inputs are annotated with an optional `Signatory`.
A `Signatory` is a trait object that can sign an individual input (given as `UnsignedTxInput`).
Added `P2PKHSignatory`, which is such a `Signatory` and can sign P2PKH inputs given the secret key, public key and sighash type.

`TxBuilder` outputs can either be `Fixed` or `Leftover`.
A `Fixed` output will always be signed as-is. A `Leftover` output can occur at most once, and it receives all the coins which are left over after deducting `Fixed` outputs and the transaction fees. It's automatically removed if it falls below the dust threshold.